Separate SolrDefault and DefaultRecord.
