+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/collection-record.phtml
