+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/data-authors.phtml
