+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/data-publicationDetails.phtml
