+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/data-summary.phtml
