+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/link-author.phtml
