+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/link-series.phtml
