+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/result-grid.phtml
